+++ full-service/src/service/api_v2.rs
@@ -1,0 +1,1 @@
+// Copyright (c) 2020-2021 MobileCoin Inc.
